chore: move updateMCDAConfig() inside MCDALayerEditor to remove circular dependency
